requirements.txt
numpy~=1.26.4
pillow~=11.1.0
datasets~=3.2.0
transformers~=4.47.1
matplotlib~=3.10.0
scikit-image~=0.25.0
accelerate>=0.26.0
evaluate~=0.4.3